Move CLineup traits
